worker/requirements.txt
amqp==5.0.1
Babel==2.5.3
billiard==3.6.3.0
celery==4.4.7
flower==0.9.5
kombu==4.6.11
pytz==2018.3
redis==3.5.3
instapy==0.6.11
instagram-scraper==1.9.0
tornado==6.0.4
vine==1.3.0
azure-storage-blob==12.5.0
numpy==1.19.3